Added code to populate Detailed results screen with details of every round played.

+++ final-year-project/StrategiesGUI/branches/SelectStrategies/src/strategiespack/Strategy.java
@@ -59,7 +59,7 @@ public class Strategy {
 		return 'n';
 	}
 	
-	public void setProbability(float prob) {
+	public void setProbability(double prob) {
 		;
 	}
